lmb-1460:  finished up the startEndDate component
