+++ app/components/mandataris/edit/form.hbs
@@ -93,7 +93,7 @@
           @endDate={{this.endDate}}
           @from={{@bestuursorgaanIT.bindingStart}}
           @to={{@bestuursorgaanIT.bindingEinde}}
-          @onUpdate={{this.updateStartEndDate}}
+          @onChange={{this.updateStartEndDate}}
           @onErrorStateUpdated={{this.updateErrorMap}}
         />
       {{/unless}}
